requirements.txt
pytorch-lightning
timm
scikit-learn
pillow
torchmetrics
pyyaml
tensorboard